This update adds price validity validation to cart drawer.
This also changes the logic of the notification and countdown - previously this notification was used only in one place on one page. But since now this block can be used everywhere - the page is used to store data and the time of the last update of the cart data.

All notifications are synchronized and show the same time.

Also, this commit update plugin version to